+++ core-ng/src/main/java/core/framework/impl/web/management/KafkaController.java
@@ -2,15 +2,23 @@ package core.framework.impl.web.management;
 
 import core.framework.impl.kafka.Kafka;
 import core.framework.impl.web.http.IPAccessControl;
+import core.framework.json.JSON;
 import core.framework.util.Lists;
+import core.framework.util.Maps;
+import core.framework.util.Strings;
 import core.framework.web.Request;
 import core.framework.web.Response;
 import org.apache.kafka.clients.admin.AdminClient;
 import org.apache.kafka.clients.admin.DescribeTopicsResult;
+import org.apache.kafka.clients.admin.NewPartitions;
+import org.apache.kafka.clients.admin.RecordsToDelete;
 import org.apache.kafka.clients.admin.TopicDescription;
 import org.apache.kafka.common.KafkaFuture;
 import org.apache.kafka.common.Node;
+import org.apache.kafka.common.TopicPartition;
 import org.apache.kafka.common.TopicPartitionInfo;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.util.List;
 import java.util.Map;
@@ -22,6 +30,7 @@ import java.util.stream.Collectors;
  * @author neo
  */
 public class KafkaController {
+    private final Logger logger = LoggerFactory.getLogger(KafkaController.class);
     private final Kafka kafka;
     private final IPAccessControl accessControl;
 
@@ -44,6 +53,40 @@ public class KafkaController {
             }
         }
         return Response.bean(views);
+    }
+
+    public Response updateTopic(Request request) {
+        String topic = request.pathParam("topic");
+        UpdateTopicRequest updateTopicRequest = request.bean(UpdateTopicRequest.class);
+        updateTopic(topic, updateTopicRequest);
+        return Response.text(Strings.format("update topic request submitted, topic={}, request={}", topic, JSON.toJSON(updateTopicRequest)));
+    }
+
+    void updateTopic(String topic, UpdateTopicRequest request) {
+        try (AdminClient admin = kafka.admin()) {
+            if (request.partitions != null) {
+                Map<String, NewPartitions> partitions = Maps.newHashMap();
+                partitions.put(topic, NewPartitions.increaseTo(request.partitions));
+                logger.info("create kafka partitions, topic={}, increaseTo={}", topic, request.partitions);
+                admin.createPartitions(partitions);
+            }
+        }
+    }
+
+    public Response deleteRecords(Request request) {
+        String topic = request.pathParam("topic");
+        DeleteRecordRequest deleteRecordRequest = request.bean(DeleteRecordRequest.class);
+        deleteRecords(topic, deleteRecordRequest);
+        return Response.text(Strings.format("delete records request submitted, topic={}, request={}", topic, JSON.toJSON(deleteRecordRequest)));
+    }
+
+    void deleteRecords(String topic, DeleteRecordRequest request) {
+        try (AdminClient admin = kafka.admin()) {
+            Map<TopicPartition, RecordsToDelete> records = Maps.newHashMap();
+            records.put(new TopicPartition(topic, request.partition), RecordsToDelete.beforeOffset(request.offset));
+            logger.info("delete kafka records, topic={}, partition={}, offset={}", topic, request.partition, request.offset);
+            admin.deleteRecords(records);
+        }
     }
 
     private KafkaTopic view(String name, TopicDescription description) {