+++ core-ng/src/main/java/core/framework/impl/web/management/UpdateTopicRequest.java
@@ -1,0 +1,11 @@
+package core.framework.impl.web.management;
+
+import core.framework.api.json.Property;
+
+/**
+ * @author neo
+ */
+public class UpdateTopicRequest {
+    @Property(name = "partitions")
+    public Integer partitions;
+}